+++ gymnax/wrappers/__init__.py
@@ -1,10 +1,10 @@
-from .brax import GymnaxToBraxWrapper
+from .dm_env import GymnaxToDmEnvWrapper
 from .evojax import GymnaxToEvoJaxTask
 from .gym import GymnaxToGymWrapper, GymnaxToVectorGymWrapper
 from .purerl import FlattenObservationWrapper, LogWrapper
 
 __all__ = [
-    "GymnaxToBraxWrapper",
+    "GymnaxToDmEnvWrapper",
     "GymnaxToEvoJaxTask",
     "GymnaxToGymWrapper",
     "GymnaxToVectorGymWrapper",